requirements.txt
natsort
pygobject # needed for pywebview
pywebview[qt]
